# Update SSHCluster usage in benchmarks with new CUDAWorker