Updates usage of `SSHCluster` according to changes in https://github.com/dask/distributed/pull/5191.