@pentschev 2021-08-12
I updated the PR and the `name` condition here @charlesbluca , could you take a look as well?